Add data warehousing.
Be sure to set AWS_ACCESS_KEY and AWS_SECRET_KEY in local_settings.py.

When navigating to /warehouse, if your group has a data warehosue (and
corresponding s3 bucket) you will have a file upload form that posts
directly to s3.  Upon completion you are redirected back you the catalog
site where you fill out the rest of the information for the resource.  A
data url is automatically created that points to your data set.